fix: 높이 너비 단위 px로 수정
