+++ src/app/(user)/mypage/page.module.css
@@ -1,0 +1,3 @@
+.container {
+  margin: 60px 20px 20px 20px;
+}
