+++ src/components/pages/MyPage/MypageForm.module.css
@@ -8,7 +8,7 @@
 
 .inputWrapper {
   flex-grow: 1;
-  height: 12%;
+  height: 35px;
 }
 
 .rowGroup {
@@ -26,6 +26,6 @@
 
 .rowGroup > .row > .firstCol {
   font-weight: 700;
-  width: 25%;
+  width: 135px;
   flex-shrink: 0;
 }
